feat(wallet): include TxSubmissionError in Wallet.transactions.outgoing.failed$
