+++ packages/wallet/src/services/index.ts
@@ -4,5 +4,4 @@ export * from './UtxoTracker';
 export * from './DelegationTracker';
 export * from './TransactionsTracker';
 export * from './AssetsTracker';
-export * from './TransactionError';
 export * from './types';
